+++ src/components/pages/schedule/ScheduleContent.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+// Import Components
+import { ViewDisplay } from './ViewDisplay';
+
+export const ScheduleContent: React.FC = () => {
+  return (
+    <>
+      <ViewDisplay />
+    </>
+  );
+};
