+++ src/pages/OptionPage.tsx
@@ -1,18 +1,13 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
-import { RootState } from '../redux';
 
 // Import Components
-import { AccountButtons, AccountSection } from '../components/pages';
+import { OptionContent } from '../components/pages';
 import { PageContainer } from '../components/shared';
 
 const OptionPage: React.FC = () => {
-  const account = useSelector((state: RootState) => state.auth.account);
-
   return (
     <PageContainer pageTitle="Settings" fullscreen>
-      {account && <AccountSection />}
-      {account && <AccountButtons />}
+      <OptionContent />
     </PageContainer>
   );
 };
